better array of iface nil entry handling for unmarshal too

+++ testdata/generic_gen_test.go
@@ -123,7 +123,7 @@ func BenchmarkDecodeCompareToIgnoredFields(b *testing.B) {
 }
 
 func TestMarshalUnmarshalMatrix(t *testing.T) {
-	v := Matrix[float64]{}
+	v := Matrix[int]{}
 	bts, err := v.MarshalMsg(nil)
 	if err != nil {
 		t.Fatal(err)
@@ -146,7 +146,7 @@ func TestMarshalUnmarshalMatrix(t *testing.T) {
 }
 
 func BenchmarkMarshalMsgMatrix(b *testing.B) {
-	v := Matrix[float64]{}
+	v := Matrix[int]{}
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
@@ -155,7 +155,7 @@ func BenchmarkMarshalMsgMatrix(b *testing.B) {
 }
 
 func BenchmarkAppendMsgMatrix(b *testing.B) {
-	v := Matrix[float64]{}
+	v := Matrix[int]{}
 	bts := make([]byte, 0, v.Msgsize())
 	bts, _ = v.MarshalMsg(bts[0:0])
 	b.SetBytes(int64(len(bts)))
@@ -167,7 +167,7 @@ func BenchmarkAppendMsgMatrix(b *testing.B) {
 }
 
 func BenchmarkUnmarshalMatrix(b *testing.B) {
-	v := Matrix[float64]{}
+	v := Matrix[int]{}
 	bts, _ := v.MarshalMsg(nil)
 	b.ReportAllocs()
 	b.SetBytes(int64(len(bts)))
@@ -181,7 +181,7 @@ func BenchmarkUnmarshalMatrix(b *testing.B) {
 }
 
 func TestEncodeDecodeMatrix(t *testing.T) {
-	v := Matrix[float64]{}
+	v := Matrix[int]{}
 	var buf bytes.Buffer
 	msgp.Encode(&buf, &v)
 
@@ -190,7 +190,7 @@ func TestEncodeDecodeMatrix(t *testing.T) {
 		t.Logf("WARNING: Msgsize() for %v is inaccurate", v)
 	}
 
-	vn := Matrix[float64]{}
+	vn := Matrix[int]{}
 	err := msgp.Decode(&buf, &vn)
 	if err != nil {
 		t.Error(err)
@@ -205,7 +205,7 @@ func TestEncodeDecodeMatrix(t *testing.T) {
 }
 
 func BenchmarkEncodeMatrix(b *testing.B) {
-	v := Matrix[float64]{}
+	v := Matrix[int]{}
 	var buf bytes.Buffer
 	msgp.Encode(&buf, &v)
 	b.SetBytes(int64(buf.Len()))
@@ -219,7 +219,7 @@ func BenchmarkEncodeMatrix(b *testing.B) {
 }
 
 func BenchmarkDecodeMatrix(b *testing.B) {
-	v := Matrix[float64]{}
+	v := Matrix[int]{}
 	var buf bytes.Buffer
 	msgp.Encode(&buf, &v)
 	b.SetBytes(int64(buf.Len()))
